+++ Resources/Prototypes/Recipes/Construction/utilities.yml
@@ -120,10 +120,10 @@
   canBuildInImpassable: false
 
 - type: construction
-  id: DisposalSignaler
+  id: DisposalSignaller
   graph: DisposalPipe
   startNode: start
-  targetNode: signaler
+  targetNode: signaller
   category: construction-category-utilities
   placementMode: SnapgridCenter
   canBuildInImpassable: false
