+++ Resources/Textures/Structures/Piping/disposal.rsi/conpipe-signaller.png
