+++ Resources/Textures/Structures/Piping/disposal.rsi/pipe-signaller.png
